De-nest relational query components
