+++ src/EntityFramework.Relational/Query/Expressions/CrossJoinExpression.cs
@@ -15,7 +15,7 @@ namespace Microsoft.Data.Entity.Relational.Query.Expressions
 
         public CrossJoinExpression([NotNull] TableExpressionBase tableExpression)
             : base(
-                Check.NotNull(tableExpression, "tableExpression").QuerySource, 
+                Check.NotNull(tableExpression, "tableExpression").QuerySource,
                 tableExpression.Alias)
         {
             _tableExpression = tableExpression;
