+++ src/EntityFramework.SQLite/Query/SQLiteQueryCompilationContext.cs
@@ -5,7 +5,9 @@ using JetBrains.Annotations;
 using Microsoft.Data.Entity.Metadata;
 using Microsoft.Data.Entity.Query;
 using Microsoft.Data.Entity.Relational.Query;
+using Microsoft.Data.Entity.Relational.Query.Methods;
 using Microsoft.Data.Entity.Relational.Query.Sql;
+using Microsoft.Data.Entity.SQLite.Utilities;
 
 namespace Microsoft.Data.Entity.SQLite.Query
 {
@@ -15,8 +17,14 @@ namespace Microsoft.Data.Entity.SQLite.Query
             [NotNull] IModel model,
             [NotNull] ILinqOperatorProvider linqOperatorProvider,
             [NotNull] IResultOperatorHandler resultOperatorHandler,
-            [NotNull] IQueryMethodProvider queryMethodProvider)
-            : base(model, linqOperatorProvider, resultOperatorHandler, queryMethodProvider)
+            [NotNull] IQueryMethodProvider queryMethodProvider,
+            [NotNull] IMethodCallTranslator methodCallTranslator)
+            : base(
+                Check.NotNull(model, "model"),
+                Check.NotNull(linqOperatorProvider, "linqOperatorProvider"),
+                Check.NotNull(resultOperatorHandler, "resultOperatorHandler"),
+                Check.NotNull(queryMethodProvider, "queryMethodProvider"),
+                Check.NotNull(methodCallTranslator, "methodCallTranslator"))
         {
         }
