+++ src/EntityFramework.SqlServer/SqlServerDataStore.cs
@@ -6,6 +6,7 @@ using Microsoft.Data.Entity.Infrastructure;
 using Microsoft.Data.Entity.Query;
 using Microsoft.Data.Entity.Relational;
 using Microsoft.Data.Entity.Relational.Query;
+using Microsoft.Data.Entity.Relational.Query.Methods;
 using Microsoft.Data.Entity.Relational.Update;
 using Microsoft.Data.Entity.SqlServer.Query;
 using Microsoft.Data.Entity.SqlServer.Update;
@@ -32,14 +33,16 @@ namespace Microsoft.Data.Entity.SqlServer
         protected override RelationalQueryCompilationContext CreateQueryCompilationContext(
             ILinqOperatorProvider linqOperatorProvider,
             IResultOperatorHandler resultOperatorHandler,
-            IQueryMethodProvider enumerableMethodProvider)
+            IQueryMethodProvider enumerableMethodProvider,
+            IMethodCallTranslator methodCallTranslator)
         {
             Check.NotNull(linqOperatorProvider, "linqOperatorProvider");
             Check.NotNull(resultOperatorHandler, "resultOperatorHandler");
             Check.NotNull(enumerableMethodProvider, "enumerableMethodProvider");
+            Check.NotNull(methodCallTranslator, "methodCallTranslator");
 
             return new SqlServerQueryCompilationContext(
-                Model, linqOperatorProvider, resultOperatorHandler, enumerableMethodProvider);
+                Model, linqOperatorProvider, resultOperatorHandler, enumerableMethodProvider, methodCallTranslator);
         }
     }
 }
