+++ src/EntityFramework/Query/EntityQueryProvider.cs
@@ -32,7 +32,7 @@ namespace Microsoft.Data.Entity.Query
 
             methodInfoBasedNodeTypeRegistry
                 .Register(AsNoTrackingExpressionNode.SupportedMethods, typeof(AsNoTrackingExpressionNode));
-            
+
             methodInfoBasedNodeTypeRegistry
                 .Register(IncludeExpressionNode.SupportedMethods, typeof(IncludeExpressionNode));
