+++ src/EntityFramework/Query/ExpressionTreeVisitors/ProjectionExpressionTreeVisitor.cs
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System;
 using System.Linq;
 using System.Linq.Expressions;
